# Conflicts:
#	centreon/cron/centreon-backup.pl
#	centreon/cron/centreon-partitioning.php
#	centreon/cron/centreon-send-stats.php
#	centreon/cron/centstorage_purge.php

+++ centreon/packages/ui/src/components/CrudPage/DeleteModal.tsx
@@ -1,11 +1,13 @@
 import { Box, CircularProgress, Typography } from '@mui/material';
+
 import { useAtom } from 'jotai';
-import { useCallback, useRef } from 'react';
+import { type ReactElement, useCallback, useRef } from 'react';
+
 import { Button } from '../Button';
 import { Modal } from '../Modal';
 import { itemToDeleteAtom } from './atoms';
 import { useDeleteItem } from './hooks/useDeleteItem';
-import { DeleteItem, ItemToDelete } from './models';
+import type { DeleteItem, ItemToDelete } from './models';
 import { isAFunction } from './utils';
 
 const DeleteModal = <TData extends { id: number; name: string }>({
@@ -30,28 +32,36 @@ const DeleteModal = <TData extends { id: number; name: string }>({
 
   const close = useCallback(() => {
     setItemToDelete(null);
-  }, []);
+  }, [setItemToDelete]);
 
   const confirm = useCallback(() => {
-    deleteItem(itemToDeleteRef.current).then(close);
-  }, [itemToDeleteRef.current]);
+    if (itemToDeleteRef.current) {
+      deleteItem(itemToDeleteRef.current).then(close);
+    }
+  }, [close, deleteItem]);
 
   if (isOpen) {
     itemToDeleteRef.current = itemToDelete;
   }
 
   return (
-    <Modal open={isOpen} onClose={close} size={modalSize}>
+    <Modal onClose={close} open={isOpen} size={modalSize}>
       <Modal.Header>
         {isAFunction(labels.title)
-          ? labels.title(itemToDeleteRef.current as TData)
-          : labels.title}
+          ? (labels.title as (item: ItemToDelete) => string | ReactElement)(
+              itemToDeleteRef.current as TData
+            )
+          : (labels.title as string | ReactElement)}
       </Modal.Header>
       <Modal.Body>
         <Typography>
           {isAFunction(labels.description)
-            ? labels.description(itemToDeleteRef.current as TData)
-            : labels.description}
+            ? (
+                labels.description as (
+                  item: ItemToDelete
+                ) => string | ReactElement
+              )(itemToDeleteRef.current as TData)
+            : (labels.description as string | ReactElement)}
         </Typography>
       </Modal.Body>
       <Box
@@ -63,10 +73,10 @@ const DeleteModal = <TData extends { id: number; name: string }>({
         }}
       >
         {isMutating && <CircularProgress size={20} />}
-        <Button variant="ghost" onClick={close} disabled={isMutating}>
+        <Button disabled={isMutating} onClick={close} variant="ghost">
           {labels.cancel}
         </Button>
-        <Button isDanger onClick={confirm} disabled={isMutating}>
+        <Button disabled={isMutating} isDanger onClick={confirm}>
           {labels.confirm}
         </Button>
       </Box>
